+++ ckb-lib-secp256k1/src/code_hashes.rs
@@ -1,1 +1,4 @@
-pub const CODE_HASH_SECP256K1: [u8; 32] = [96, 254, 141, 45, 108, 149, 203, 10, 61, 193, 107, 16, 177, 23, 164, 120, 139, 10, 139, 29, 169, 219, 126, 44, 116, 13, 244, 29, 234, 237, 116, 231];
+pub const CODE_HASH_SECP256K1: [u8; 32] = [
+    96, 254, 141, 45, 108, 149, 203, 10, 61, 193, 107, 16, 177, 23, 164, 120, 139, 10, 139, 29,
+    169, 219, 126, 44, 116, 13, 244, 29, 234, 237, 116, 231,
+];
