+++ tests/src/tests/secp256k1_blake2b_sighash_all.rs
@@ -1,29 +1,18 @@
+use crate::helper::{blake160, sign_secp256k1_blake2b_sighash_all};
+
 use super::*;
-use crate::helper::{blake160, sign_tx};
 
 use ckb_system_scripts::BUNDLED_CELL;
-use ckb_testtool::ckb_crypto::secp::{Generator, Privkey};
-use ckb_testtool::ckb_error::Error;
+use ckb_testtool::ckb_crypto::secp::Generator;
 use ckb_testtool::ckb_types::{bytes::Bytes, core::TransactionBuilder, packed::*, prelude::*};
 use ckb_testtool::context::Context;
 
+use std::fs;
+
 const MAX_CYCLES: u64 = 10_000_000;
 
-// error numbers
-const ERROR_EMPTY_ARGS: i8 = 2;
-
-fn assert_script_error(err: Error, err_code: i8) {
-    let error_string = err.to_string();
-    assert!(
-        error_string.contains(format!("error code {} ", err_code).as_str()),
-        "error_string: {}, expected_error_code: {}",
-        error_string,
-        err_code
-    );
-}
-
 #[test]
-fn test_success() {
+fn test_secp256k1_blake2b_sighash_all() {
     // generate key pair
     let privkey = Generator::random_privkey();
     let pubkey = privkey.pubkey().expect("pubkey");
@@ -91,60 +80,11 @@ fn test_success() {
     let tx = context.complete_tx(tx);
 
     // sign
-    let tx = sign_tx(tx, &privkey);
+    let tx = sign_secp256k1_blake2b_sighash_all(tx, &privkey);
 
     // run
     let cycles = context
         .verify_tx(&tx, MAX_CYCLES)
         .expect("pass verification");
     println!("consume cycles: {}", cycles);
-}
-
-#[test]
-fn test_empty_args() {
-    // deploy contract
-    let mut context = Context::default();
-    let contract_bin: Bytes = Loader::default().load_binary("otx-lock");
-    let out_point = context.deploy_cell(contract_bin);
-
-    // prepare scripts
-    let lock_script = context
-        .build_script(&out_point, Default::default())
-        .expect("script");
-
-    // prepare cells
-    let input_out_point = context.create_cell(
-        CellOutput::new_builder()
-            .capacity(1000u64.pack())
-            .lock(lock_script.clone())
-            .build(),
-        Bytes::new(),
-    );
-    let input = CellInput::new_builder()
-        .previous_output(input_out_point)
-        .build();
-    let outputs = vec![
-        CellOutput::new_builder()
-            .capacity(500u64.pack())
-            .lock(lock_script.clone())
-            .build(),
-        CellOutput::new_builder()
-            .capacity(500u64.pack())
-            .lock(lock_script)
-            .build(),
-    ];
-
-    let outputs_data = vec![Bytes::new(); 2];
-
-    // build transaction
-    let tx = TransactionBuilder::default()
-        .input(input)
-        .outputs(outputs)
-        .outputs_data(outputs_data.pack())
-        .build();
-    let tx = context.complete_tx(tx);
-
-    // run
-    let err = context.verify_tx(&tx, MAX_CYCLES).unwrap_err();
-    assert_script_error(err, ERROR_EMPTY_ARGS);
 }
